fixes test
